+++ src/lib/components/DateRangePicker/index.ts
@@ -1,5 +1,2 @@
-import DateRangePicker from './DateRangePicker.svelte'
-
-export { DateRangePicker }
-export * from './date-range-picker.types.js'
-export * from './date-range-picker.variants.js'
+export { default as DateRangePicker } from './DateRangePicker.svelte'
+export type { DateRangePickerProps } from './date-range-picker.types.js'