+++ src/lib/components/InputNumber/index.ts
@@ -1,4 +1,2 @@
 export { default as InputNumber } from './InputNumber.svelte'
-export type { InputNumberProps } from './inputNumber.types.js'
-export { inputNumberVariants, inputNumberDefaults } from './inputNumber.variants.js'
-export type { InputNumberVariantProps, InputNumberSlots } from './inputNumber.variants.js'
+export type { InputNumberProps } from './input-number.types.js'