+++ src/lib/components/TagsInput/index.ts
@@ -1,5 +1,0 @@
-import TagsInput from './TagsInput.svelte'
-
-export { TagsInput }
-export * from './tags-input.types.js'
-export * from './tags-input.variants.js'